requirements.txt
Flask
pandas
faiss-cpu
numpy
sentence-transformers
transformers